Refactor touch into its own component Fix Doxygen warnings, header layout Update IDF links
