+++ attic/components/hdw-spiffs/include/hdw-spiffs.h
@@ -4,7 +4,7 @@
  *
  * SPIFFS is a file system intended for SPI NOR flash devices on embedded targets. It supports wear levelling, file
  * system consistency checks, and more. The full API reference can be found here: <a
- * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.5/esp32s2/api-reference/storage/spiffs.html">SPIFFS
+ * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.7/esp32s2/api-reference/storage/spiffs.html">SPIFFS
  * Filesystem</a>.
  *
  * Ths Swadge treats SPIFFS as a read-only file system.
