+++ components/hdw-battmon/include/hdw-battmon.h
@@ -6,11 +6,11 @@
  * \section battmon_design Design Philosophy
  *
  * The battery monitor uses the <a
- * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.5/esp32s2/api-reference/peripherals/adc_oneshot.html">Analog
+ * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.7/esp32s2/api-reference/peripherals/adc_oneshot.html">Analog
  * to Digital Converter (ADC) Oneshot Mode Driver</a>.
  *
  * The battery monitor code is based on the <a
- * href="https://github.com/espressif/esp-idf/tree/v5.2.5/examples/peripherals/adc/oneshot_read">ADC Single Read
+ * href="https://github.com/espressif/esp-idf/tree/v5.2.7/examples/peripherals/adc/oneshot_read">ADC Single Read
  * Example</a>.
  *
  * \warning The battery monitor and microphone (hdw-mic.h) cannot be used at the same time! Each mode can either
