+++ components/hdw-dac/include/hdw-dac.h
@@ -3,7 +3,7 @@
  * \section dac_design Design Philosophy
  *
  * The Digital To Analog Converter (DAC) is used to drive a speaker or headphones. It uses the <a
- * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.5/esp32s2/api-reference/peripherals/dac.html">DAC
+ * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.7/esp32s2/api-reference/peripherals/dac.html">DAC
  * peripheral</a>. The peripheral can be used a few different ways, but this project uses DMA to output a
  * continuous, arbitrary signal.
  *
