+++ components/hdw-touch/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "hdw-touch.c"
+                    INCLUDE_DIRS "include"
+                    REQUIRES driver esp_timer)
