+++ components/hdw-usb/include/hdw-usb.h
@@ -3,9 +3,9 @@
  * \section usb_design Design Philosophy
  *
  * The USB component uses Espressif's <a
- * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.5/esp32s2/api-reference/peripherals/usb_device.html">USB
+ * href="https://docs.espressif.com/projects/esp-idf/en/v5.2.7/esp32s2/api-reference/peripherals/usb_device.html">USB
  * Device Driver</a>. It's based on the <a
- * href="https://github.com/espressif/esp-idf/tree/v5.2.5/examples/peripherals/usb/device/tusb_hid">TinyUSB
+ * href="https://github.com/espressif/esp-idf/tree/v5.2.7/examples/peripherals/usb/device/tusb_hid">TinyUSB
  * Human Interface Device Example</a>.
  *
  * The Swadge primarily functions as a USB gamepad.
