+++ emulator/src/components/hdw-btn/hdw-btn_emu.h
@@ -6,6 +6,5 @@
 #include "hdw-btn.h"
 
 void emulatorInjectButton(buttonBit_t button, bool down);
-void emulatorSetTouchJoystick(int32_t phi, int32_t radius, int32_t intensity);
 void emulatorHandleKeys(int keycode, int bDown);
 buttonBit_t emulatorGetButtonState(void);
