+++ emulator/src/components/hdw-touch/hdw-touch_emu.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <stdint.h>
+#include <stdbool.h>
+
+#include "hdw-touch.h"
+
+void emulatorSetTouchJoystick(int32_t phi, int32_t radius, int32_t intensity);
