+++ emulator/src/emu_main.c
@@ -31,6 +31,8 @@
 #include "hdw-led_emu.h"
 #include "hdw-btn.h"
 #include "hdw-btn_emu.h"
+#include "hdw-touch.h"
+#include "hdw-touch_emu.h"
 #include "hdw-imu_emu.h"
 
 #include "hdw-mic.h"
