+++ emulator/src/extensions/gamepad/ext_gamepad.c
@@ -2,6 +2,7 @@
 
 #include "emu_utils.h"
 #include "hdw-btn_emu.h"
+#include "hdw-touch_emu.h"
 #include "hdw-imu_emu.h"
 #include "macros.h"
 #include "trigonometry.h"
