+++ main/swadge2024.h
@@ -180,6 +180,7 @@
 #include "hdw-imu.h"
 #include "hdw-battmon.h"
 #include "hdw-btn.h"
+#include "hdw-touch.h"
 #include "hdw-ch32v003.h"
 #include "hdw-dac.h"
 #include "hdw-esp-now.h"
